# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Sub-Registry API
  description: Enterprise Model Context Protocol (MCP) sub-registry for server discovery and governance
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@company.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3010
    description: Development server
  - url: https://mcp-registry.company.com
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
...
paths:
...
  /v0/servers:
    get:
      tags: [Servers]
      summary: List MCP servers
      description: Retrieve a paginated list of MCP servers with optional filtering
      security:
        - {}
        - ApiKeyAuth: [read]
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of servers to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of servers to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          description: Filter servers by status
          schema:
            $ref: '#/components/schemas/ServerStatus'
        - name: search
          in: query
          description: Search servers by name or description
          schema:
            type: string
      responses:
        '200':
          description: List of servers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v0/servers/{id}:
    get:
      tags: [Servers]
      summary: Get server by ID
      description: Retrieve detailed information about a specific MCP server
      security:
        - {}
        - ApiKeyAuth: [read]
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Server unique identifier
          schema:
            type: string
      responses:
        '200':
          description: Server details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPServer'
        '404':
          description: Server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v0/publish:
    post:
      tags: [Servers]
      summary: Publish MCP server
      description: Register a new MCP server in the registry
      security:
        - ApiKeyAuth: [publish]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishServerRequest'
      responses:
        '201':
          description: Server published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPServer'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Server name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key authentication using "ApiKey <key>" format
  schemas:
...
    MCPServer:
      type: object
      required: [id, name, description, version, status]
      properties:
        id:
          type: string
          description: Unique server identifier
        name:
          type: string
          description: Server name in reverse DNS format
          example: com.company.mcp-server
        description:
          type: string
          description: Human-readable server description
        version:
          type: string
          description: Semantic version
          example: 1.0.0
        status:
          $ref: '#/components/schemas/ServerStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        repository:
          $ref: '#/components/schemas/Repository'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        remote:
          $ref: '#/components/schemas/RemoteConfig'
        metadata:
          type: object
          additionalProperties: true
          description: Additional server metadata
    ServerStatus:
      type: string
      enum: [experimental, beta, stable, deprecated]
      description: Server development and deployment status
    Repository:
      type: object
      required: [type, url]
      properties:
        type:
          type: string
          enum: [git, mercurial, svn]
        url:
          type: string
          format: uri
        branch:
          type: string
        tag:
          type: string
        commit:
          type: string
    Package:
      type: object
      required: [registry, identifier, version]
      properties:
        registry:
          type: string
          enum: [npm, pypi, maven, docker, cargo, gem]
        identifier:
          type: string
          description: Package identifier in the registry
        version:
          type: string
          description: Package version
        url:
          type: string
          format: uri
          description: Direct download URL
    RemoteConfig:
      type: object
      required: [transport]
      properties:
        transport:
          type: string
          enum: [stdio, http, https, tcp, websocket]
        url:
          type: string
          format: uri
        host:
          type: string
        port:
          type: integer
        path:
          type: string

    # Request/Response Types
    PublishServerRequest:
      type: object
      required: [name, description, version, status]
      properties:
        name:
          type: string
          pattern: '^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)*$'
          description: Server name in reverse DNS format
        description:
          type: string
          minLength: 10
          maxLength: 500
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+(-[\w\d\-_]+)?(\+[\w\d\-_]+)?$'
          description: Semantic version
        status:
          $ref: '#/components/schemas/ServerStatus'
        repository:
          $ref: '#/components/schemas/Repository'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        remote:
          $ref: '#/components/schemas/RemoteConfig'
        metadata:
          type: object
          additionalProperties: true
    ServerListResponse:
      type: object
      required: [servers, pagination]
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/MCPServer'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        has_more:
          type: boolean

    # Authentication Types
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        timestamp:
          type: string
          format: date-time
    ValidationError:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Validation error message
        code:
          type: string
          description: Error code (typically VALIDATION_ERROR)
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string